+++ Editor/Addressables/AddressableGroupRules.cs
@@ -12,7 +12,7 @@ namespace UnityEditor.Localization.Addressables
     /// Provides support for configuring rule sets to determine the <see cref="AddressableAssetGroup"/>s that localized assets should be added to.
     /// </summary>
     /// <example>
-    /// This example places all English assets into a local group and all other languages into a remote group which could then be downloaded after the game is released.
+    /// <para>This example places all English assets into a local group and all other languages into a remote group which could then be downloaded after the game is released.</para>
     /// <code source="../../DocCodeSamples.Tests/GroupResolverExample.cs"/>
     /// </example>
     public class AddressableGroupRules : ScriptableObject